requirements.txt
recipe-scrapers==15.8.0
ingredient-parser-nlp==2.1.1
anthropic==0.57.1
openai>=1.0.0
ollama>=0.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
beautifulsoup4>=4.12.0
requests>=2.28.0
jinja2>=3.0.0